requirements.txt
langchain_core
langchain
python-dotenv
langchain_openai
langchain_google_genai
google-generativeai
